backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
google-cloud-compute==1.19.0
pydantic==2.12.5
pydantic-core==2.41.5
pytest==8.0.0
python-dotenv==1.0.1
pytest-asyncio==0.23.6
